Add initial docs, and some thoughts about final interfaces
